requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytest==8.3.4
pytest-mock==3.11.1
pytest-asyncio==0.25.3
motor==3.7.0
pymongo==4.11.1 